Refactor EndpointSliceFieldIndexes
