+++ internal/manager/index/endpointslice.go
@@ -15,10 +15,17 @@ const (
 	KubernetesServiceNameLabel = "kubernetes.io/service-name"
 )
 
-// ServiceNameIndexFunc is a client.IndexerFunc that parses a Kubernetes object and returns the value of the
+// CreateEndpointSliceFieldIndices creates a FieldIndices map for the EndpointSlice resource.
+func CreateEndpointSliceFieldIndices() FieldIndices {
+	return FieldIndices{
+		KubernetesServiceNameIndexField: serviceNameIndexFunc,
+	}
+}
+
+// serviceNameIndexFunc is a client.IndexerFunc that parses a Kubernetes object and returns the value of the
 // Kubernetes service-name label.
 // Used to index EndpointSlices by their service owners.
-func ServiceNameIndexFunc(obj client.Object) []string {
+func serviceNameIndexFunc(obj client.Object) []string {
 	slice, ok := obj.(*discoveryV1.EndpointSlice)
 	if !ok {
 		panic(fmt.Sprintf("expected an EndpointSlice; got %T", obj))
