+++ internal/manager/index/index.go
@@ -1,0 +1,6 @@
+package index
+
+import "sigs.k8s.io/controller-runtime/pkg/client"
+
+// FieldIndices is a map of field names to their indexer functions.
+type FieldIndices map[string]client.IndexerFunc
